+++ src/entities/study/curriculum/model/curriculum.queries.types.ts
