+++ src/entities/study/notice/model/notice.queries.types.ts
@@ -1,0 +1,9 @@
+export interface UsePostNoticeProps {
+  onSuccess?: () => void;
+  onError?: (error: unknown) => void;
+  studyId: number
+}
+
+export interface usePachNoticeProps extends UsePostNoticeProps {
+  noticeId: number;
+}
