--- src/features/notice/fetch-notice/model/constants.ts
@@ -1,5 +1,0 @@
-// notice qery key
-export const NOTICE_QUERY_KEY = {
-  all: (studyId: number) => ["getNotices", studyId],
-  detail: (studyId: number, noticeId: number) => ["getNotices", studyId, noticeId],
-}
