+++ src/features/study/curriculum/add-curriculum/model/useAddCurriculum.ts
@@ -1,6 +1,6 @@
 import { useFormContext } from "react-hook-form";
 import { useState } from "react";
-import { FormValues } from "@/entities/study/curriculum/model/curriculum.types";
+import { FormValues } from "@/entities/study/curriculum/model/curriculum.queries.types";
 
 const useAddCurriculum = () => {
   // 커리큘럼 추가 버튼 플래그
