+++ src/features/study/curriculum/delete-curriculum/model/useDeleteCurriculum.ts
@@ -1,4 +1,4 @@
-import { FormValues } from "@/entities/study/curriculum/model/curriculum.types";
+import { FormValues } from "@/entities/study/curriculum/model/curriculum.queries.types";
 import { useFormContext } from "react-hook-form";
 
 const useDeleteCurriculum = () => {
